[HotFix] Fix missing argument

+++ custom_functions/custom_loss.py
@@ -18,8 +18,8 @@ class MeanFalseError(object):
         y_pred = ops.convert_to_tensor(y_pred)
         y_true = math_ops.cast(y_true, y_pred.dtype)
 
-        n_positive = math_ops.cast(tf_count(y_true, 1.0))
-        n_negative = math_ops.cast(tf_count(y_true, 0.0))
+        n_positive = math_ops.cast(tf_count(y_true, 1.0), tf.float32)
+        n_negative = math_ops.cast(tf_count(y_true, 0.0), tf.float32)
 
         positive_indices = tf.where(condition = tf.equal(y_true, 1.0))
         negative_indices = tf.where(condition = tf.equal(y_true, 0.0))
@@ -40,8 +40,8 @@ class MeanFalseError(object):
         y_pred = ops.convert_to_tensor(y_pred)
         y_true = math_ops.cast(y_true, y_pred.dtype)
 
-        n_positive = math_ops.cast(tf_count(y_true, 1.0))
-        n_negative = math_ops.cast(tf_count(y_true, 0.0))
+        n_positive = math_ops.cast(tf_count(y_true, 1.0), tf.float32)
+        n_negative = math_ops.cast(tf_count(y_true, 0.0), tf.float32)
 
         positive_indices = tf.where(condition = tf.equal(y_true, 1.0))
         negative_indices = tf.where(condition = tf.equal(y_true, 0.0))
